unify all to a single command, refactor out the ChatAttachPromptAction
